+++ flutter_quill_extensions/lib/logic/extensions/attribute.dart
@@ -11,4 +11,20 @@ class MobileHeightAttribute extends Attribute<String?> {
       : super('mobileHeight', AttributeScope.ignore, val);
 }
 
-extension AttributeExt on Attribute {}
+class MobileMarginAttribute extends Attribute<String?> {
+  const MobileMarginAttribute(String? val)
+      : super('mobileMargin', AttributeScope.ignore, val);
+}
+
+class MobileAlignmentAttribute extends Attribute<String?> {
+  const MobileAlignmentAttribute(String? val)
+      : super('mobileAlignment', AttributeScope.ignore, val);
+}
+
+extension AttributeExt on Attribute {
+  static const MobileWidthAttribute mobileWidth = MobileWidthAttribute(null);
+  static const MobileHeightAttribute mobileHeight = MobileHeightAttribute(null);
+  static const MobileMarginAttribute mobileMargin = MobileMarginAttribute(null);
+  static const MobileAlignmentAttribute mobileAlignment =
+      MobileAlignmentAttribute(null);
+}